+++ vector/src/main/java/im/vector/fragments/VectorSettingsPreferencesFragment.kt
@@ -28,7 +28,6 @@ import android.net.Uri
 import android.os.AsyncTask
 import android.os.Build
 import android.os.Bundle
-import android.os.Parcelable
 import android.provider.Settings
 import android.support.design.widget.TextInputEditText
 import android.support.v14.preference.SwitchPreference
@@ -56,7 +55,10 @@ import im.vector.activity.*
 import im.vector.contacts.ContactsManager
 import im.vector.extensions.getFingerprintHumanReadable
 import im.vector.extensions.withArgs
-import im.vector.preference.*
+import im.vector.preference.ProgressBarPreference
+import im.vector.preference.UserAvatarPreference
+import im.vector.preference.VectorGroupPreference
+import im.vector.preference.VectorPreference
 import im.vector.settings.FontScale
 import im.vector.settings.VectorLocale
 import im.vector.ui.themes.ThemeUtils
@@ -119,6 +121,8 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
     private var mMyDeviceInfo: DeviceInfo? = null
 
     private var mDisplayedPushers = ArrayList<Pusher>()
+
+    private var interactionListener: VectorSettingsFragmentInteractionListener? = null
 
     // devices: device IDs and device names
     private var mDevicesNameList: List<DeviceInfo> = ArrayList()
@@ -205,9 +209,6 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
     }
     private val backgroundSyncPreference by lazy {
         findPreference(PreferencesManager.SETTINGS_ENABLE_BACKGROUND_SYNC_PREFERENCE_KEY) as SwitchPreference
-    }
-    private val mRingtonePreference by lazy {
-        findPreference(PreferencesManager.SETTINGS_NOTIFICATION_RINGTONE_SELECTION_PREFERENCE_KEY)
     }
     private val mUseRiotCallRingtonePreference by lazy {
         findPreference(PreferencesManager.SETTINGS_CALL_RINGTONE_USE_RIOT_PREFERENCE_KEY) as SwitchPreference
@@ -396,21 +397,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
         }
         refreshNotificationPrivacy()
 
-        // Ringtone
-        mRingtonePreference.onPreferenceClickListener = Preference.OnPreferenceClickListener {
-            val intent = Intent(RingtoneManager.ACTION_RINGTONE_PICKER)
-            intent.putExtra(RingtoneManager.EXTRA_RINGTONE_TYPE, RingtoneManager.TYPE_NOTIFICATION)
-
-            if (null != PreferencesManager.getNotificationRingTone(activity)) {
-                intent.putExtra(RingtoneManager.EXTRA_RINGTONE_EXISTING_URI, PreferencesManager.getNotificationRingTone(activity))
-            }
-
-            startActivityForResult(intent, REQUEST_NOTIFICATION_RINGTONE)
-            false
-        }
-        refreshNotificationRingTone()
-
-        for (resourceText in mPushesRuleByResourceId.keys) {
+        for (resourceText in mPrefKeyToBingRuleId.keys) {
             val preference = findPreference(resourceText)
 
             if (null != preference) {
@@ -420,40 +407,6 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
                         // the callback is called even if there is no user interaction
                         // so the value will be checked to ensure there is really no update.
                         onPushRuleClick(preference.key, newValueAsVoid as Boolean)
-                        true
-                    }
-                } else if (preference is BingRulePreference) {
-                    preference.onPreferenceClickListener = Preference.OnPreferenceClickListener {
-                        AlertDialog.Builder(activity!!)
-                                .setSingleChoiceItems(R.array.notification_status, preference.ruleStatusIndex) { d, index ->
-                                    val rule = preference.createRule(index)
-                                    d.cancel()
-
-                                    if (null != rule) {
-                                        displayLoadingView()
-                                        mSession.dataHandler
-                                                .bingRulesManager
-                                                .updateRule(preference.rule,
-                                                        rule,
-                                                        object : BingRulesManager.onBingRuleUpdateListener {
-                                                            private fun onDone() {
-                                                                refreshDisplay()
-                                                                hideLoadingView()
-                                                            }
-
-                                                            override fun onBingRuleUpdateSuccess() {
-                                                                onDone()
-                                                            }
-
-                                                            override fun onBingRuleUpdateFailure(errorMessage: String) {
-                                                                activity?.toast(errorMessage)
-
-                                                                onDone()
-                                                            }
-                                                        })
-                                    }
-                                }
-                                .show()
                         true
                     }
                 }
@@ -867,6 +820,18 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
         }
     }
 
+    override fun onAttach(context: Context?) {
+        super.onAttach(context)
+        if (context is VectorSettingsFragmentInteractionListener) {
+            interactionListener = context
+        }
+    }
+
+    override fun onDetach() {
+        interactionListener = null
+        super.onDetach()
+    }
+
     override fun onResume() {
         super.onResume()
 
@@ -907,6 +872,12 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
             refreshNotificationPrivacy()
             refreshDisplay()
             refreshBackgroundSyncPrefs()
+        }
+
+        interactionListener?.requestedKeyToHighlight()?.let { key ->
+            interactionListener?.requestHighlightPreferenceKeyOnResume(null)
+            val preference = findPreference(key)
+            (preference as? VectorPreference)?.isHighlighted = true
         }
     }
 
@@ -1007,16 +978,11 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
         val pushManager = Matrix.getInstance(appContext)!!.pushManager
 
-        for (resourceText in mPushesRuleByResourceId.keys) {
+        for (resourceText in mPrefKeyToBingRuleId.keys) {
             val preference = preferenceManager.findPreference(resourceText)
 
             if (null != preference) {
-                if (preference is BingRulePreference) {
-                    preference.isEnabled = null != rules && isConnected && pushManager.areDeviceNotificationsAllowed()
-                    mSession.dataHandler.pushRules()?.let {
-                        preference.setBingRule(it.findDefaultRule(mPushesRuleByResourceId[resourceText]))
-                    }
-                } else if (preference is SwitchPreference) {
+                if (preference is SwitchPreference) {
                     if (resourceText == PreferencesManager.SETTINGS_ENABLE_THIS_DEVICE_PREFERENCE_KEY) {
                         preference.isChecked = pushManager.areDeviceNotificationsAllowed()
                     } else if (resourceText == PreferencesManager.SETTINGS_TURN_SCREEN_ON_PREFERENCE_KEY) {
@@ -1033,8 +999,6 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
         // If notifications are disabled for the current user account or for the current user device
         // The others notifications settings have to be disable too
         val areNotificationAllowed = rules?.findDefaultRule(BingRule.RULE_ID_DISABLE_ALL)?.isEnabled == true
-
-        mRingtonePreference.isEnabled = !areNotificationAllowed && pushManager.areDeviceNotificationsAllowed()
 
         mNotificationPrivacyPreference.isEnabled = !areNotificationAllowed && pushManager.areDeviceNotificationsAllowed() && pushManager.useFcm()
     }
@@ -1200,7 +1164,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
             return
         }
 
-        val ruleId = mPushesRuleByResourceId[fResourceText]
+        val ruleId = mPrefKeyToBingRuleId[fResourceText]
         val rule = mSession.dataHandler.pushRules()?.findDefaultRule(ruleId)
 
         // check if there is an update
@@ -1306,13 +1270,6 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
     }
 
     /**
-     * Refresh the notification ring tone
-     */
-    private fun refreshNotificationRingTone() {
-        mRingtonePreference.summary = PreferencesManager.getNotificationRingToneName(activity)
-    }
-
-    /**
      * Refresh the notification privacy setting
      */
     private fun refreshNotificationPrivacy() {
@@ -1333,17 +1290,6 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
         if (resultCode == Activity.RESULT_OK) {
             when (requestCode) {
-                REQUEST_NOTIFICATION_RINGTONE -> {
-                    PreferencesManager.setNotificationRingTone(activity,
-                            data?.getParcelableExtra<Parcelable>(RingtoneManager.EXTRA_RINGTONE_PICKED_URI) as Uri?)
-
-                    // test if the selected ring tone can be played
-                    if (null == PreferencesManager.getNotificationRingToneName(activity)) {
-                        PreferencesManager.setNotificationRingTone(activity, PreferencesManager.getNotificationRingTone(activity))
-                    }
-
-                    refreshNotificationRingTone()
-                }
                 REQUEST_CALL_RINGTONE -> {
                     val callRingtoneUri: Uri? = data?.getParcelableExtra(RingtoneManager.EXTRA_RINGTONE_PICKED_URI)
                     if (callRingtoneUri != null) {
@@ -1419,11 +1365,11 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
             putString(PreferencesManager.SETTINGS_VERSION_PREFERENCE_KEY, VectorUtils.getApplicationVersion(activity))
 
             mSession.dataHandler.pushRules()?.let {
-                for (resourceText in mPushesRuleByResourceId.keys) {
+                for (resourceText in mPrefKeyToBingRuleId.keys) {
                     val preference = findPreference(resourceText)
 
                     if (null != preference && preference is SwitchPreference) {
-                        val ruleId = mPushesRuleByResourceId[resourceText]
+                        val ruleId = mPrefKeyToBingRuleId[resourceText]
 
                         val rule = it.findDefaultRule(ruleId)
                         var isEnabled = null != rule && rule.isEnabled
@@ -2851,21 +2797,13 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
         private const val REQUEST_NEW_PHONE_NUMBER = 456
         private const val REQUEST_PHONEBOOK_COUNTRY = 789
         private const val REQUEST_LOCALE = 777
-        private const val REQUEST_NOTIFICATION_RINGTONE = 888
         private const val REQUEST_CALL_RINGTONE = 999
 
-        // rule Id <-> preference name
-        private var mPushesRuleByResourceId = mapOf(
+        // preference name <-> rule Id
+        private var mPrefKeyToBingRuleId = mapOf(
                 PreferencesManager.SETTINGS_ENABLE_ALL_NOTIF_PREFERENCE_KEY to BingRule.RULE_ID_DISABLE_ALL,
                 PreferencesManager.SETTINGS_ENABLE_THIS_DEVICE_PREFERENCE_KEY to DUMMY_RULE,
-                PreferencesManager.SETTINGS_TURN_SCREEN_ON_PREFERENCE_KEY to DUMMY_RULE,
-                PreferencesManager.SETTINGS_CONTAINING_MY_DISPLAY_NAME_PREFERENCE_KEY to BingRule.RULE_ID_CONTAIN_DISPLAY_NAME,
-                PreferencesManager.SETTINGS_CONTAINING_MY_USER_NAME_PREFERENCE_KEY to BingRule.RULE_ID_CONTAIN_USER_NAME,
-                PreferencesManager.SETTINGS_MESSAGES_IN_ONE_TO_ONE_PREFERENCE_KEY to BingRule.RULE_ID_ONE_TO_ONE_ROOM,
-                PreferencesManager.SETTINGS_MESSAGES_IN_GROUP_CHAT_PREFERENCE_KEY to BingRule.RULE_ID_ALL_OTHER_MESSAGES_ROOMS,
-                PreferencesManager.SETTINGS_INVITED_TO_ROOM_PREFERENCE_KEY to BingRule.RULE_ID_INVITE_ME,
-                PreferencesManager.SETTINGS_CALL_INVITATIONS_PREFERENCE_KEY to BingRule.RULE_ID_CALL,
-                PreferencesManager.SETTINGS_MESSAGES_SENT_BY_BOT_PREFERENCE_KEY to BingRule.RULE_ID_SUPPRESS_BOTS_NOTIFICATIONS
+                PreferencesManager.SETTINGS_TURN_SCREEN_ON_PREFERENCE_KEY to DUMMY_RULE
         )
 
         // static constructor